Special H&E stain division
Pre-specifies a DIV in the HTML image-container for H&E ("HE") stain
slides, to ensure that these slides are shown first. Requested by our
pathologists, may not be of general interest. Note that a corresponding
change in viewer.js, setting and slides with tag "H&E" to instead have
tag "HE" for consistency, was included in the previous change.